pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinelscope"
version = "0.1.0"
description = "Advanced attack surface recon and security reporting toolkit"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "SentinelScope contributors" }]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "typer>=0.12.3",
  "httpx>=0.27.0",
  "dnspython>=2.6.1",
  "cryptography>=43.0.0",
  "jinja2>=3.1.4",
  "rich>=13.7.1",
  "pydantic>=2.7.4",
  "requests>=2.32.3",
]

[project.scripts]
sscan = "sentinelscope.cli:main"

[tool.setuptools]
package-dir = {"" = "."}
include-package-data = true

[tool.setuptools.package-data]
"sentinelscope.reporting" = ["templates/*.html"]

[tool.setuptools.packages.find]
include = ["sentinelscope*"]

